+++ include/calloc.h
@@ -1,0 +1,10 @@
+#include "include/memory.h"
+
+/**
+ * @brief Asigna un bloque de memoria para un número de elementos, inicializándolo a cero.
+ *
+ * @param number Número de elementos.
+ * @param size Tamaño de cada elemento.
+ * @return void* Puntero al área de datos asignada e inicializada.
+ */
+void* calloc(size_t number, size_t size);
