+++ include/malloc.h
@@ -1,0 +1,35 @@
+#include "include/memory.h"
+
+/**
+ * @brief Encuentra un bloque libre que tenga al menos el tamaño solicitado.
+ *
+ * @param last Puntero al último bloque.
+ * @param size Tamaño solicitado.
+ * @return t_block Puntero al bloque encontrado, o NULL si no se encuentra ninguno.
+ */
+t_block find_block(t_block* last, size_t size);
+
+/**
+ * @brief Divide un bloque de memoria en dos, si el tamaño solicitado es menor que el bloque disponible.
+ *
+ * @param b Bloque a dividir.
+ * @param s Tamaño del nuevo bloque.
+ */
+void split_block(t_block b, size_t s);
+
+/**
+ * @brief Expande el heap para crear un nuevo bloque de memoria.
+ *
+ * @param last Último bloque del heap.
+ * @param s Tamaño del nuevo bloque.
+ * @return t_block Puntero al nuevo bloque creado.
+ */
+t_block extend_heap(t_block last, size_t s);
+
+/**
+ * @brief Asigna un bloque de memoria del tamaño solicitado.
+ *
+ * @param size Tamaño en bytes del bloque a asignar.
+ * @return void* Puntero al área de datos asignada.
+ */
+void* malloc(size_t size);
