+++ include/realloc.h
@@ -1,0 +1,18 @@
+#include "include/memory.h"
+
+/**
+ * @brief Copia el contenido de un bloque de origen a un bloque de destino.
+ *
+ * @param src Bloque de origen.
+ * @param dst Bloque de destino.
+ */
+void copy_block(t_block src, t_block dst);
+
+/**
+ * @brief Cambia el tamaño de un bloque de memoria previamente asignado.
+ *
+ * @param p Puntero al área de datos a redimensionar.
+ * @param size Nuevo tamaño en bytes.
+ * @return void* Puntero al área de datos redimensionada.
+ */
+void* realloc(void* p, size_t size);
